requirements.txt
playwright
python-dotenv
openai
pydantic
ipython
tabulate
pandas
streamlit